Added support for restic --verify flag
Added support for restic volume checksum verification.
This adds another pod annotation ('backup.velero.io/verify-volumes')
which is used to identify which volumes on the pod should get the
'--verify' flag passed to restic when restoring.

Stdout and stderr are logged to the restic pod logs. The PodVolumeRestore
CRD has three new status fields: Errors, VerifyErrors, and ResticPod,
which contain the number of total errors reported, the number of verify
errors reported, and the name of the restic pod that ran the restore
(to allow for the user to look in the pod logs for more detailed output).

The Restore CRD has two new Status fields: PodVolumeRestoreErrors
and PodVolumeRestoreVerifyErrors, which contain a slice of ObjectReferences
listing which PodVolumeRestores contained errors or verify errors.

There is not yet support built into restore describe to see the restore fields
via the velero client. Also, the restic daemonset will need to have an env
variable set for POD_NAME.

(cherry picked from commit da2df66ad156ccd3c86dfb885634d744abc1b980)